+++ pages/admin.vue
@@ -2,73 +2,12 @@
 import Toast from "@/components/Toast";
 import LoadingAdmin from "@/components/LoadingAdmin.vue";
 import { useStatusStore } from "@/stores/useStatus.js";
-
-// cart Store
-const statusStore = useStatusStore();
-
-// layout default false
+import { useUser } from "@/composables/useUser";
 definePageMeta({
 	layout: false,
 });
-
-// 取 .env
-const config = useRuntimeConfig();
-
-// toast store
-import { useToastStore } from "@/stores/useToast.js";
-const store = useToastStore();
-
-// router
-const router = useRouter();
-
-const token = ref("");
-const checkSuccess = ref(false);
-const checkLogin = () => {
-	token.value = document.cookie.replace(
-		/(?:(?:^|.*;\s*)token\s*=\s*([^;]*).*$)|^.*$/,
-		"$1"
-	);
-	const api = `${config.public.apiUrl}/auth/check`;
-	$fetch(api, {
-		method: "POST",
-		headers: {
-			"Content-Type": "application/json",
-			Accept: "application/json",
-			Authorization: `Bearer ${token.value}`,
-		},
-		body: {
-			api_token: token.value,
-		},
-	})
-		.then((res) => {
-			checkSuccess.value = true;
-		})
-		.catch(() => {
-			router.push("/login");
-		});
-};
-
-const signout = () => {
-	const api = `${config.public.apiUrl}/auth/logout`;
-
-	$fetch(api, {
-		method: "POST",
-		headers: {
-			"Content-Type": "application/json",
-			Accept: "application/json",
-			Authorization: `Bearer ${token.value}`,
-		},
-		body: {
-			api_token: token.value,
-		},
-	}).then((res) => {
-		store.messageHandle(res.message);
-		store.isShowHandle();
-		document.cookie = "token=; expires=; path=/";
-		router.push("/");
-	});
-};
-
+const statusStore = useStatusStore();
+const { checkSuccess, checkLogin, signOut } = useUser();
 onMounted(() => {
 	checkLogin();
 });
@@ -92,7 +31,7 @@ onMounted(() => {
 						<NuxtLink to="/" class="nav-link active">回首頁</NuxtLink>
 					</li>
 					<li class="nav-item text-nowrap">
-						<a href="#" class="nav-link" @click.prevent="signout">登出</a>
+						<a href="#" class="nav-link" @click.prevent="signOut">登出</a>
 					</li>
 				</ul>
 			</div>
